Move test stubs to domain
