+++ src/main/tv/codely/scala_http_api/entry_point/Routes.scala
@@ -27,6 +27,6 @@ final class Routes(container: EntryPointDependencyContainer) {
       }
     }
 
-  private def jsonBody[T](handler: Map[String, JsValue] => Route): Route =
+  private def jsonBody(handler: Map[String, JsValue] => Route): Route =
     entity(as[JsValue])(json => handler(json.asJsObject.fields))
 }
