--- src/test/tv/codely/scala_http_api/module/shared/stub/DurationStub.scala
@@ -1,7 +1,0 @@
-package tv.codely.scala_http_api.module.shared.stub
-
-import scala.concurrent.duration._
-
-object DurationStub {
-  def random: Duration = IntStub.randomUnsigned().minutes
-}
