--- src/test/tv/codely/scala_http_api/module/shared/stub/IntStub.scala
@@ -1,11 +1,0 @@
-package tv.codely.scala_http_api.module.shared.stub
-
-import scala.util.Random
-
-object IntStub {
-  def random: Int = Random.nextInt()
-
-  def randomUnsigned(max: Int = Int.MaxValue - 1): Int = randomBetween(min = 0, max = max)
-
-  def randomBetween(min: Int, max: Int): Int = Random.nextInt((max - min) + 1)
-}
