--- src/test/tv/codely/scala_http_api/module/shared/stub/StringStub.scala
@@ -1,7 +1,0 @@
-package tv.codely.scala_http_api.module.shared.stub
-
-import scala.util.Random
-
-object StringStub {
-  def random(numChars: Int): String = Random.alphanumeric take numChars mkString ""
-}
