--- src/test/tv/codely/scala_http_api/module/shared/stub/UuidStub.scala
@@ -1,9 +1,0 @@
-package tv.codely.scala_http_api.module.shared.stub
-
-import java.util.UUID
-
-object UuidStub {
-  def apply(value: String): UUID = UUID.fromString(value)
-
-  def random: UUID = UUID.randomUUID()
-}
