+++ src/test/tv/codely/scala_http_api/module/user/domain/UserIdStub.scala
@@ -1,9 +1,8 @@
-package tv.codely.scala_http_api.module.user.infrastructure.stub
+package tv.codely.scala_http_api.module.user.domain
 
 import java.util.UUID
 
-import tv.codely.scala_http_api.module.shared.stub.UuidStub
-import tv.codely.scala_http_api.module.user.domain.UserId
+import tv.codely.scala_http_api.module.shared.domain.UuidStub
 
 object UserIdStub {
   def apply(value: String): UserId = UserIdStub(UuidStub(value))
