+++ src/test/tv/codely/scala_http_api/module/user/domain/UserNameStub.scala
@@ -1,7 +1,6 @@
-package tv.codely.scala_http_api.module.user.infrastructure.stub
+package tv.codely.scala_http_api.module.user.domain
 
-import tv.codely.scala_http_api.module.shared.stub.{IntStub, StringStub}
-import tv.codely.scala_http_api.module.user.domain.UserName
+import tv.codely.scala_http_api.module.shared.domain.{IntStub, StringStub}
 
 object UserNameStub {
   private val minimumChars = 1
