+++ src/test/tv/codely/scala_http_api/module/video/domain/VideoCategoryStub.scala
@@ -1,7 +1,6 @@
-package tv.codely.scala_http_api.module.video.infrastructure.stub
+package tv.codely.scala_http_api.module.video.domain
 
-import tv.codely.scala_http_api.module.shared.stub.IntStub
-import tv.codely.scala_http_api.module.video.domain.VideoCategory
+import tv.codely.scala_http_api.module.shared.domain.IntStub
 
 object VideoCategoryStub {
   private val categories = Seq("Screencast", "Interviews")
