+++ src/test/tv/codely/scala_http_api/module/video/domain/VideoIdStub.scala
@@ -1,10 +1,9 @@
-package tv.codely.scala_http_api.module.video.infrastructure.stub
+package tv.codely.scala_http_api.module.video.domain
 
 import java.util.UUID
 
-import tv.codely.scala_http_api.module.shared.stub.UuidStub
+import tv.codely.scala_http_api.module.shared.domain.UuidStub
 import tv.codely.scala_http_api.module.user.domain.UserId
-import tv.codely.scala_http_api.module.video.domain.VideoId
 
 object VideoIdStub {
   def apply(value: String): VideoId = VideoIdStub(UuidStub(value))
