+++ src/test/tv/codely/scala_http_api/module/video/domain/VideoTitleStub.scala
@@ -1,7 +1,6 @@
-package tv.codely.scala_http_api.module.video.infrastructure.stub
+package tv.codely.scala_http_api.module.video.domain
 
-import tv.codely.scala_http_api.module.shared.stub.{IntStub, StringStub}
-import tv.codely.scala_http_api.module.video.domain.VideoTitle
+import tv.codely.scala_http_api.module.shared.domain.{IntStub, StringStub}
 
 object VideoTitleStub {
   private val minimumChars = 1
